# docs/openapi.yaml
components:
   schemas:
      Body_create_product_products__post:
         properties:
            category_names:
               title: Category Names
               type: string
            description:
               title: Description
               type: string
            image:
               format: binary
               title: Image
               type: string
            name:
               title: Name
               type: string
            price:
               title: Price
               type: integer
         required:
         - name
         - description
         - category_names
         - price
         - image
         title: Body_create_product_products__post
         type: object
      Body_update_image_media__image_id__put:
         properties:
            image:
               format: binary
               title: Image
               type: string
         required:
         - image
         title: Body_update_image_media__image_id__put
         type: object
      Body_update_product_products__product_id__put:
         properties:
            category_names:
               title: Category Names
               type: string
            description:
               title: Description
               type: string
            name:
               title: Name
               type: string
            price:
               title: Price
               type: integer
         required:
         - name
         - description
         - category_names
         - price
         title: Body_update_product_products__product_id__put
         type: object
      Category:
         properties:
            description:
               anyOf:
               -  type: string
               -  type: 'null'
               title: Description
            id:
               title: Id
               type: integer
            is_active:
               default: true
               title: Is Active
               type: boolean
            name:
               title: Name
               type: string
         required:
         - id
         - name
         title: Category
         type: object
...
      HTTPValidationError:
         properties:
            detail:
               items:
                  $ref: '#/components/schemas/ValidationError'
               title: Detail
               type: array
         title: HTTPValidationError
         type: object
      ImageWithProduct:
         properties:
            id:
               title: Id
               type: integer
            path:
               title: Path
               type: string
            product:
               $ref: '#/components/schemas/Product'
            product_id:
               title: Product Id
               type: integer
         required:
         - id
         - product_id
         - path
         - product
         title: ImageWithProduct
         type: object
      Product:
         properties:
            categories:
               default: []
               items:
                  $ref: '#/components/schemas/Category'
               title: Categories
               type: array
            category_names:
               description: "\u0412\u043E\u0437\u0432\u0440\u0430\u0449\u0430\u0435\
                  \u0442 \u043D\u0430\u0437\u0432\u0430\u043D\u0438\u044F \u043A\u0430\
                  \u0442\u0435\u0433\u043E\u0440\u0438\u0439 \u0434\u043B\u044F \u0443\
                  \u0434\u043E\u0431\u0441\u0442\u0432\u0430"
               items:
                  type: string
               readOnly: true
               title: Category Names
               type: array
            description:
               title: Description
               type: string
            id:
               title: Id
               type: integer
            name:
               title: Name
               type: string
            price:
               title: Price
               type: integer
         required:
         - id
         - name
         - description
         - price
         - category_names
         title: Product
         type: object
      ValidationError:
         properties:
            loc:
               items:
                  anyOf:
                  -  type: string
                  -  type: integer
               title: Location
               type: array
            msg:
               title: Message
               type: string
            type:
               title: Error Type
               type: string
         required:
         - loc
         - msg
         - type
         title: ValidationError
         type: object
info:
   description: "\u0420\u0435\u0444\u0430\u043A\u0442\u043E\u0440\u0438\u0440\u043E\
      \u0432\u0430\u043D\u043D\u044B\u0439 API \u0434\u043B\u044F \u0441\u0438\u0441\
      \u0442\u0435\u043C\u044B \u0443\u043F\u0440\u0430\u0432\u043B\u0435\u043D\u0438\
      \u044F \u043A\u043E\u043B\u043B\u0430\u0436\u0430\u043C\u0438"
   title: CollagrCms API
   version: 0.1.0
openapi: 3.1.0
paths:
...
   /media/{image_id}:
      get:
         description: "\u0412\u043E\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442\
            \ \u0444\u0430\u0439\u043B \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\
            \u043D\u0438\u044F \u043F\u043E ID\n\nArgs:\n    image_id: ID \u0438\u0437\
            \u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u044F\n\nReturns:\n \
            \   \u0424\u0430\u0439\u043B \u0438\u0437\u043E\u0431\u0440\u0430\u0436\
            \u0435\u043D\u0438\u044F"
         operationId: get_image_media__image_id__get
         parameters:
         -  in: path
            name: image_id
            required: true
            schema:
               title: Image Id
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Get Image
         tags:
         - images
      put:
         description: "\u041E\u0431\u043D\u043E\u0432\u043B\u044F\u0435\u0442 \u0438\
            \u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u0435\n\nArgs:\n\
            \    image_id: ID \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\
            \u0438\u044F\n    image: \u041D\u043E\u0432\u044B\u0439 \u0444\u0430\u0439\
            \u043B \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u044F\
            \n\nReturns:\n    \u0421\u043E\u043E\u0431\u0449\u0435\u043D\u0438\u0435\
            \ \u043E\u0431 \u0443\u0441\u043F\u0435\u0448\u043D\u043E\u043C \u043E\
            \u0431\u043D\u043E\u0432\u043B\u0435\u043D\u0438\u0438"
         operationId: update_image_media__image_id__put
         parameters:
         -  in: path
            name: image_id
            required: true
            schema:
               title: Image Id
               type: integer
         requestBody:
            content:
               multipart/form-data:
                  schema:
                     $ref: '#/components/schemas/Body_update_image_media__image_id__put'
            required: true
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Update Image
         tags:
         - images
   /products/:
      get:
         description: "\u041F\u043E\u043B\u0443\u0447\u0430\u0435\u0442 \u0432\u0441\
            \u0435 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u044B \u0441 \u0438\u0437\
            \u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u044F\u043C\u0438\n\n\
            Args:\n    request: HTTP \u0437\u0430\u043F\u0440\u043E\u0441 \u0434\u043B\
            \u044F \u043F\u043E\u043B\u0443\u0447\u0435\u043D\u0438\u044F base_url\n\
            \nReturns:\n    \u0421\u043F\u0438\u0441\u043E\u043A \u043F\u0440\u043E\
            \u0434\u0443\u043A\u0442\u043E\u0432 \u0441 \u0438\u0437\u043E\u0431\u0440\
            \u0430\u0436\u0435\u043D\u0438\u044F\u043C\u0438"
         operationId: "\u043F\u043E\u043B\u0443\u0447\u0438\u0442\u044C_\u043F\u0440\
            \u043E\u0434\u0443\u043A\u0442\u044B_products__get"
         responses:
            '200':
               content:
                  application/json:
                     schema:
                        items:
                           $ref: '#/components/schemas/ImageWithProduct'
                        title: "Response \u041F\u043E\u043B\u0443\u0447\u0438\u0442\
                           \u044C \u041F\u0440\u043E\u0434\u0443\u043A\u0442\u044B\
                           \ Products  Get"
                        type: array
               description: Successful Response
         summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u041F\u0440\u043E\
            \u0434\u0443\u043A\u0442\u044B"
         tags:
         - products
      post:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u043D\u043E\u0432\
            \u044B\u0439 \u043F\u0440\u043E\u0434\u0443\u043A\u0442 \u0441 \u0438\u0437\
            \u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u0435\u043C\n\nArgs:\n\
            \    name: \u041D\u0430\u0437\u0432\u0430\u043D\u0438\u0435 \u043F\u0440\
            \u043E\u0434\u0443\u043A\u0442\u0430\n    description: \u041E\u043F\u0438\
            \u0441\u0430\u043D\u0438\u0435 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\
            \u0430\n    category_names: \u041D\u0430\u0437\u0432\u0430\u043D\u0438\
            \u044F \u043A\u0430\u0442\u0435\u0433\u043E\u0440\u0438\u0439 \u0447\u0435\
            \u0440\u0435\u0437 \u0437\u0430\u043F\u044F\u0442\u0443\u044E\n    price:\
            \ \u0426\u0435\u043D\u0430 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u0430\
            \n    image: \u0424\u0430\u0439\u043B \u0438\u0437\u043E\u0431\u0440\u0430\
            \u0436\u0435\u043D\u0438\u044F\n\nReturns:\n    \u0418\u043D\u0444\u043E\
            \u0440\u043C\u0430\u0446\u0438\u044F \u043E \u0441\u043E\u0437\u0434\u0430\
            \u043D\u043D\u043E\u043C \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u0435\
            \ \u0438 \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u0438"
         operationId: create_product_products__post
         requestBody:
            content:
               multipart/form-data:
                  schema:
                     $ref: '#/components/schemas/Body_create_product_products__post'
            required: true
         responses:
            '201':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create Product
         tags:
         - products
   /products/by-category/{category_name}/:
      get:
         description: "\u041F\u043E\u043B\u0443\u0447\u0430\u0435\u0442 \u0432\u0441\
            \u0435 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u044B \u043E\u043F\u0440\
            \u0435\u0434\u0435\u043B\u0435\u043D\u043D\u043E\u0439 \u043A\u0430\u0442\
            \u0435\u0433\u043E\u0440\u0438\u0438\n\nArgs:\n    category_name: \u041D\
            \u0430\u0437\u0432\u0430\u043D\u0438\u0435 \u043A\u0430\u0442\u0435\u0433\
            \u043E\u0440\u0438\u0438\n    request: HTTP \u0437\u0430\u043F\u0440\u043E\
            \u0441 \u0434\u043B\u044F \u043F\u043E\u043B\u0443\u0447\u0435\u043D\u0438\
            \u044F base_url\n\nReturns:\n    \u0421\u043F\u0438\u0441\u043E\u043A\
            \ \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u043E\u0432 \u0443\u043A\u0430\
            \u0437\u0430\u043D\u043D\u043E\u0439 \u043A\u0430\u0442\u0435\u0433\u043E\
            \u0440\u0438\u0438"
         operationId: get_product_category_products_by_category__category_name___get
         parameters:
         -  in: path
            name: category_name
            required: true
            schema:
               title: Category Name
               type: string
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Get Product Category
         tags:
         - products
...
   /products/{product_id}:
      delete:
         description: "\u0423\u0434\u0430\u043B\u044F\u0435\u0442 \u043F\u0440\u043E\
            \u0434\u0443\u043A\u0442 \u0438 \u0441\u0432\u044F\u0437\u0430\u043D\u043D\
            \u043E\u0435 \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\
            \u0435\n\nArgs:\n    product_id: ID \u043F\u0440\u043E\u0434\u0443\u043A\
            \u0442\u0430\n\nReturns:\n    \u0421\u043E\u043E\u0431\u0449\u0435\u043D\
            \u0438\u0435 \u043E\u0431 \u0443\u0441\u043F\u0435\u0448\u043D\u043E\u043C\
            \ \u0443\u0434\u0430\u043B\u0435\u043D\u0438\u0438"
         operationId: delete_product_products__product_id__delete
         parameters:
         -  in: path
            name: product_id
            required: true
            schema:
               title: Product Id
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Delete Product
         tags:
         - products
      put:
         description: "\u041E\u0431\u043D\u043E\u0432\u043B\u044F\u0435\u0442 \u043F\
            \u0440\u043E\u0434\u0443\u043A\u0442\n\nArgs:\n    product_id: ID \u043F\
            \u0440\u043E\u0434\u0443\u043A\u0442\u0430\n    name: \u041D\u043E\u0432\
            \u043E\u0435 \u043D\u0430\u0437\u0432\u0430\u043D\u0438\u0435 \u043F\u0440\
            \u043E\u0434\u0443\u043A\u0442\u0430\n    description: \u041D\u043E\u0432\
            \u043E\u0435 \u043E\u043F\u0438\u0441\u0430\u043D\u0438\u0435 \u043F\u0440\
            \u043E\u0434\u0443\u043A\u0442\u0430\n    category_names: \u041D\u0430\
            \u0437\u0432\u0430\u043D\u0438\u044F \u043A\u0430\u0442\u0435\u0433\u043E\
            \u0440\u0438\u0439 \u0447\u0435\u0440\u0435\u0437 \u0437\u0430\u043F\u044F\
            \u0442\u0443\u044E\n    price: \u041D\u043E\u0432\u0430\u044F \u0446\u0435\
            \u043D\u0430 \u043F\u0440\u043E\u0434\u0443\u043A\u0442\u0430\n\nReturns:\n\
            \    \u0421\u043E\u043E\u0431\u0449\u0435\u043D\u0438\u0435 \u043E\u0431\
            \ \u0443\u0441\u043F\u0435\u0448\u043D\u043E\u043C \u043E\u0431\u043D\u043E\
            \u0432\u043B\u0435\u043D\u0438\u0438"
         operationId: update_product_products__product_id__put
         parameters:
         -  in: path
            name: product_id
            required: true
            schema:
               title: Product Id
               type: integer
         requestBody:
            content:
               application/x-www-form-urlencoded:
                  schema:
                     $ref: '#/components/schemas/Body_update_product_products__product_id__put'
            required: true
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Update Product
         tags:
         - products